Allow partial import of reports
Before: if some of the reports sent in a request failed to import for
some reason, none were imported (and proxy deleted them all).

Now: all reports that can be imported can be imported, errors are
logged. If at least one report is imported successfully, proxy deletes
all of them.